Add python unit tests and integration tests
This commit adds three python tests where we check the get, set features
for dragon vectors. Also an integration test where  we train the model
only using the compressed dragon gradients. Increased the learning rate
to 0.002 instead of 0.001.

The model is trained only on the compressed weight gradients, biases are
left as it is.