Add integration tests for the compression interface
This commit adds tests namely get set tests and a training
test. For some reason, the  training test in dag api gives
a much lower accuracy whereas the same test gives a high
accuracy in boltpython. The get set functions are working
fine which means there is a problem with updating the gradi-
-ents somewhere.