Allow ProgressBar when console_logging is disabled
Not sure if this is a good idea, but the idea is to construct the
progress bar if and only if the verbose thingy is on, and in addition,
there is no logging enabled to console.

Otherwise console-logging (to stderr) and ProgressBar updates (to
stdout) can mess up the display.