Disable thirdai.distributed_bolt import with package (#739)
Apple M1 Silicon ends up with an irrecoverrable error on a default install with ray.
Trying to import GRPC fails. To avoid this from happening, this commit changes
the package to no longer import thirdai.distributed_bolt by default. Users are 
expected to explicitly import the package as thirdai.distributed_bolt. 

Also includes fix for test-PyPI workflows to be tested in main. 

Reviewed-By: Jerin Philip <jerin@thirdai.com>