Fix PyPI workflow on GitHub CI (#744)
Older versions of python (3.6) on cibuildwheel are failing due to import
differences, breaking the CI workflow for wheels.  This change fixes this with
an `import thirdai.distributed`. This gets us further in the build on CI but
Windows builds appear to be failing due to lack of `ray` package via PyPI.
Testing ray is temporarily disabled on Windows to get the wheels workflow green
(as Windows distributed is not high priority).

See also: #641, #739
Reviewed-By: Pratik Pranav <pratik@thirdai.com>