Etch version to library (py, cpp) from single source of truth (#746)
We want C++, Python (and future bindings) if any to be able to report versions
at runtime, so we can checkout to the particular commit for debugging or
reproducing bugs purposes. We should be able to hook this onto mlflow, logging
etc so a trail is left when we have to investigate something or go back in time
to test something. The additions should also help reproducibility.

The reported string is intended to be set as:

Format:   [Major].[Minor].[Patch]+[Commit-short-SHA]
Example:  0.1.8+c838f9e

[Major], [Minor], [Patch] are read from a `thirdai.version` file in the root
folder (Not using VERSION because the naming has caused issues with MacOS and
WebAssembly before). setup.py, which previously had the value hardcoded and was
bumped from time to time is reconfigured to read from `thirdai.version`.
[Commit-short-SHA] is obtained by Git package in CMake and configured into a
header file at build time.