Clang support (#653)
This change gets `cmake ..` work out of the box on MacOS with AppleClang.

Additionally there are a few issues in code which clang brings to attention,
and are fixed while achieving clang compilation. Most of these issues are regarding
placement of explicit template instantiations, which have to be now moved down
to the bottom of files after the definitions appear.

Modern CMake suggests using target OpenMP::OpenMP_CXX to get -fopenmp (gcc)
-fopenmp=libomp (for clang). Using this target makes OpenMP compilation errors
go away.