Build only _thirdai (no tests/all) via setup.py (#777)
There is no need to wait for all the test executables to build while the
package is being built via setup.py. From what I see locally, there is a long
waiting for the test executables to build for a change that's only related to
python (no testing, >2 min). Towards simplifying the situation, the bare
minimum requirement (i.e _thirdai, which is the pybind target) alone is
configured to be built via setup.py (while the wheel builds are in play).

The time gain from not having to build a lot of executables should also reflect
across CI and wheel builds.