Add concat to the compression interface (#797)
* Add concat to the compression interface
This commit exposes concat feature to python. You can
now make a list of compressed vectors in python and
send them to C++ to concatenate them. A test has also been
added to test_dragon to test concat. I do not think so we
need to write a seperate test for count sketches as we only
need to test the data transfer from c++ to python and back.

* Fix nits

* Fix Nits