Serializer compression (#760)
* Add serialize method and deserialize method

Change int seed for hashing to uint32

* Fix Serializer

This commit fixes the serializer. We now have a constructor  for the
dragon vector class that accepts a stringstream. No function called
deserialize now.

* Add serializer to CountSketch

This commit adds a serializer to the count sketch class.
-fopenmp issues in this commit at the moment.

* Remove uint32 from header, Remove readBinaryString func

This commit removes readBinaryString function from Compression
Utils.h file

* Add serialization size

* Add serialize size to count sketch

* Add py::bytes to std::string conversion
Starting with this commit, Parameter reference can now
accept a py::byte object and make a compressed vector
from the byte object. There is an implicit type-casting
from py::bytes to std::String which we have to profile. This
type casting makes use of a copy function.

* Initialize bytes object directly from stringstream

* Add concat to the compression interface (#797)

* Add concat to the compression interface
This commit exposes concat feature to python. You can
now make a list of compressed vectors in python and
send them to C++ to concatenate them. A test has also been
added to test_dragon to test concat. I do not think so we
need to write a seperate test for count sketches as we only
need to test the data transfer from c++ to python and back.

* Fix nits

* Fix Nits

* Remove dynamic casting horror from CompressionFactory
concat in compression factory looks much simpler now
there is a pure virtual function in the Compressed Vector
class called extend that accepts a unique pointer to
a Compressed Vector object. The implementation of ext
end in derived classes uses dynamic casting. So now
dynamic dispatch for concat looks kinda slick. :D

* Add serialize to an array
Up till now, we had been working with stringstreams to ser-
ialize data. This is kind of inefficient because stringstreams
have to own the underlying string. Hence, there are copies
involved. Also, to transfer stringstreams to python, we have to
first get the string from the stringstream which is a copy oper-
ation and then further converted into python byte array or what-
ever. We can rather work with char arrays. Since, char is a much
more primitive datatype and we have more control over how it is
tranferred and constructed, we can make things more efficient. I
am thinking of transferring the ownership of this char array
to python. This way, we wouldn't make any copies or anything.

* Add constructor dragon vector
This constructor can take in a char array and deserialize it
into a dragon vector

* Add char array serializer to Count Sketch
Count sketch also has a char array serializer now,
and added a corresponding constructor for the same.
ToDo: Converting char  array to python object without
copying

* Remove py bytes serialized logic
A lot of things in this commit. Now, char arrays work end
to end with python.

* Test count sketch test, Remove Print statements

* Fixing pragma

* Plant an enum into the serialized compressed vector.

* Solve nits
Move using compressed vector to a more global scope inside
the parameter reference class.
Remove useless headers.
Remove print statements from python tests file

* Remove multiple copies of  same compressed objects
Deleted copy costructor and then fixed places wherever an error
was being thrown. But there is still a copy constructor at the time
being.

* Adding comments

* Change CompressedVector to FloatCompressedVector
make_unique in CompressionFactory is simpler now
Extend methods in dragonvector and count sketch are also
simpler now. Made a few naming changes

* Add cstring to header files

* Change test function names

* Add comment to concat in compression factory

* Fix serialized_size, earlier we were overapproximating it