Logging improvements (#846)
Exposes more of the logging module to python thirdai.logging, with
documentation to work with sphinx. The following functions are now exposed:

    # Setup now brought under thirdai.logging module.
    # (previously thirdai.setupLogging)
    thirdai.logging.setup(...) 

    # Ability to write to C++ logs from Python.
    thirdai.logging.trace(...)
    thirdai.logging.debug(...)
    thirdai.logging.info(...)
    thirdai.logging.warn(...)
    thirdai.logging.error(...)
    thirdai.logging.critical(...)

    # Flush capability
    thirdai.logging.flush()

Work in distributed and some dogfooding has brought about some issues with log
flushing, which should be improved by this PR. In case someone needs to force a
flush, there's a function now exposed to Python to do the same. A default flush
interval of 10s is applied.

There's also an option to log-every n-updates so the updates are less busy.
Default is one update.

* Logging module refactor
* Refactor public API
  * Also logging::setupLogging -> logging::setup(...)
  * Make logging default pattern more parseable

* Logging miscallaneous enhancements
  * Log train loss only at specified intervals of updates
  * Account for setup re-entry with different args
  * Flush every n-seconds; 
  * Expose flush(); 
  * Use shutdown() instead of drop_all()
    
BoltGraph::train - epoch, updates consolidatation:
  * Consolidate epoch, _epoch_cnt -> _epoch; _batch_cnt -> _updates
  * Move train_state.epoch = _epoch near increment

LR Scheduler tests: 
  * Epochs are now [1, 3] when num_epochs=4 coz 0-based indexing