Add mid-batch validation capability (#847)
It's important to have validation set and validation loss as part of training,
from time to time. This change wires an option to training to validate every n
updates, provided a validation set during training. As of now it's only
possible to validate at the end of an epoch. This is troublesome for problems
like MLM where there is no periodic feedback on whether the metrics are
improving on the unseen validation set or not, and only training loss is
visible.

The validation losses are logged in addition to training to give quicker
real-time feedback to the user and can be later on parsed to produced training
curves.

Minor: Fixing =0 case for log-loss-frequency as well.

See also: TSK-469
Reviewed-By: Yashwanth Adunukota  <yash@thirdai.com>