Save best model per tracked validation metric (#891)
Adds an option to save the best model per validation loss. The best model is as
of now saved as {model_prefix}.best.bolt and the most recent one saved as
{model_prefix}.last.bolt.

Documentation updated to reflect. Some consolidation in Metric.h to share
across EarlyStopCheckpoint callback and dropping validation checkpoints.

Logs with level info are dropped when models are saved onto disk, to provide
some feedback on the location of models written.

See: TSK-471