Move large definitions from headers to source-files (#985)
.cc files are more amenable to compilation speed gains using tools like
ccache - which our source code is getting ripe for. This change moves a
bunch of definitions currently in headers in bolt-core to source-files.
A large chunk of code here is move of .h in the core library to .cc
files.  Since instrumentation has revealed that template work from
cereal is a source of compilation slowdowns, this PR also attempts to
move some cereal related things into .cc files. This is done using a
follow the compiler error and fix approach.

Doing explicit instantiations require operating similar to cereal's docs
on the pimpl idiom.